pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "sae_auto_interp"
version = "0.1.0"
description = "Automated Interpretability"
readme = "README.md"
requires-python = ">=3.10"
keywords = ["interpretability", "explainable-ai"]
dependencies = [
    "datasets",
    "nnsight",
    "orjson",
    "sae@git+https://github.com/EleutherAI/sae@4c73890dfe94529037fa5d0f2592124f80f41f56",
    "safetensors",
    "simple_parsing",
    "torchtyping",
    "fire",
    "blobfile",
    "transformer_lens",
    "bitsandbytes",
    "flask",
    "vllm",
    "aiofiles",
    "sentence_transformers",
    "pip>=25.0.1",
    "ipykernel>=6.29.5",
    "armaan@git+https://github.com/armaan-abraham/tools.git",
    "mlsae@git+https://github.com/armaan-abraham/mlsae@f814b288b7041c334f083615ab8ec388681be9e7",
    "ipywidgets>=8.1.5",
    "seaborn>=0.13.2",
    "pandas>=2.2.3",
    "statannotations>=0.7.1",
]

[tool.pyright]
include = ["sae_auto_interp*"]
reportPrivateImportUsage = false

[tool.setuptools.packages.find]
include = ["sae_auto_interp*"]

[tool.ruff]
# TODO: Clean up or remove experiments folder.
exclude = ["experiments", "**/prompts/*", "**/prompts.py","**/simulator.py"]
# Avoid automatically removing unused imports in __init__.py files.
# Such imports will be flagged with a dedicated message suggesting
# that the import is either added to the module's __all__ symbol
ignore-init-module-imports = true
# Ignore F821 (undefined name) errors, as they are common in type annotations.
ignore = ["F722", "F821"]
# Same as Black.
line-length = 88
# Enable pycodestyle (`E`), Pyflakes (`F`), and isort (`I`) codes
# See https://beta.ruff.rs/docs/rules/ for more possible rules
select = ["E", "F", "I"]
